==> analysis-survey/01-index-new-deltas.Rmd <==
---
title: "Spatiotemporal models of `r paste(params$species)` biomass with RW"
author: "Philina English"
date: "`r Sys.Date()`"
output: html_document
params:
    # species: "Canary Rockfish"
    # species: "Arrowtooth Flounder"
    # species: "Big Skate"
    # species: "North Pacific Spiny Dogfish"
    # species: "Yelloweye Rockfish"
    species: "Yellowtail Rockfish"
    # species: "Redstripe Rockfish"
    # species: "Pacific Ocean Perch"
    # species: "Bocaccio"
    # species: "Splitnose Rockfish"
    #species: "Pacific Sand Lance"
    # species: "Longspine Thornyhead"
    # species: "Sandpaper Skate"
    region: "Coast-wide trawl surveys"
    # region: "HBLL outside surveys"
    # region: "HBLL inside surveys"
    # region: "HBLL north outside survey"
    # region: "HBLL south outside survey"
    # region: "WCHG only"
    name: "-rw-no-covs" # covariate string used to name model rds
    silent: FALSE
---

```{r global_options, include=FALSE}
dir.create("data-generated", showWarnings = FALSE)
knitr::opts_chunk$set(
  fig.width = 11, fig.height = 8.5,
  echo = FALSE, warning = FALSE, message = FALSE
)
library(dplyr)
library(ggplot2)
library(gfplot)
library(gfdata)
library(sdmTMB)

theme_set(
  gfplot::theme_pbs(base_size = 14)
)

```

```{r params}
species <- params$species
region <- params$region
name <- params$name

paste("region =", region)
paste("model label =", name)
```
Load data
```{r, message=FALSE, warning=FALSE}
spp <- gsub(" ", "-", gsub("\\/", "-", tolower(species)))

if (region == "Both odd year trawl surveys") {
  nd_all <- readRDS(here::here("data-generated/synoptic-grid.rds"))
  survey <- c("SYN QCS", "SYN HS")
  type <- "Synoptic trawl"
  model_ssid <- c(1, 3)
}

if (region == "QCS only") {
  nd_all <- readRDS(here::here("data-generated/synoptic-grid.rds"))
  survey <- c("SYN QCS")
  type <- "Synoptic trawl"
  model_ssid <- c(1)
}

if (region == "HS only") {
  nd_all <- readRDS(here::here("data-generated/synoptic-grid.rds"))
  survey <- c("SYN HS")
  type <- "Synoptic trawl"
  model_ssid <- c(3)
}

if (region == "WCVI only") {
  nd_all <- readRDS(here::here("data-generated/synoptic-grid.rds"))
  survey <- c("SYN WCVI")
  type <- "Synoptic trawl"
  model_ssid <- c(4)
}

if (region == "WCHG only") {
  nd_all <- readRDS(here::here("data-generated/synoptic-grid.rds"))
  survey <- c("SYN WCHG")
  type <- "Synoptic trawl"
  model_ssid <- c(16)
}

if (region == "Coast-wide trawl surveys") {
  nd_all <- readRDS(here::here("data-generated/synoptic-grid.rds"))
  survey <- c("SYN QCS", "SYN HS", "SYN WCVI", "SYN WCHG")
  type <- "Synoptic trawl"
  model_ssid <- c(1, 3, 4, 16)
}

if (region == "North (PMFC 5DE)") {
  nd_all <- readRDS(here::here("data-generated/synoptic-grid.rds"))
  nd_5D <- filter(nd_all, survey == "SYN HS") %>% filter(Y > 5907.814)
  nd_5E <- filter(nd_all, survey == "SYN WCHG")
  nd_all <- bind_rows(nd_5D, nd_5E)
  survey <- c("SYN HS", "SYN WCHG")
  type <- "Synoptic trawl"
  model_ssid <- c(3, 16)
}

if (region == "South (PMFC 3CD5ABC)") {
  nd_all <- readRDS(here::here("data-generated/synoptic-grid.rds"))
  nd_5C <- filter(nd_all, survey == "SYN HS") %>% filter(Y < 5907.814)
  nd_south <- filter(nd_all, survey %in% c("SYN QCS", "SYN WCVI"))
  nd_all <- bind_rows(nd_5C, nd_south)
  survey <- c("SYN HS", "SYN QCS", "SYN WCVI")
  type <- "Synoptic trawl"
  model_ssid <- c(1, 3, 4)
}

if (region == "South 5 only (PMFC 5ABC)") {
  nd_all <- readRDS(here::here("data-generated/synoptic-grid.rds"))
  nd_5C <- filter(nd_all, survey == "SYN HS") %>% filter(Y < 5907.814)
  nd_QCS <- filter(nd_all, survey %in% c("SYN QCS"))
  nd_all <- bind_rows(nd_5C, nd_QCS)
  survey <- c("SYN HS", "SYN QCS")
  type <- "Synoptic trawl"
  model_ssid <- c(1, 3)
}

if (region == "HS & WCHG") {
  nd_all <- readRDS(here::here("data-generated/synoptic-grid.rds"))
  survey <- c("SYN HS", "SYN WCHG")
  type <- "Synoptic trawl"
  model_ssid <- c(3, 16)
}

if (region == "QCS & WCVI") {
  nd_all <- readRDS(here::here("data-generated/synoptic-grid.rds"))
  survey <- c("SYN QCS", "SYN WCVI")
  type <- "Synoptic trawl"
  model_ssid <- c(1, 4)
}

if (region == "HBLL outside surveys") {
  nd_all <- readRDS(here::here("data-generated/hbll-outside-grid.rds"))
  survey <- c("HBLL OUT N", "HBLL OUT S")
  type <- "HBLL (outside)"
  model_ssid <- c(22, 36)
}

if (region == "HBLL south outside survey") {
  nd_all <- readRDS(here::here("data-generated/hbll-s-outside-grid.rds"))
  survey <- c("HBLL OUT S")
  type <- "HBLL (outside)"
  model_ssid <- c(36)
}

if (region == "HBLL north outside survey") {
  nd_all <- readRDS(here::here("data-generated/hbll-n-outside-grid.rds"))
  survey <- c("HBLL OUT N")
  type <- "HBLL (outside)"
  model_ssid <- c(22)
}

if (region == "HBLL inside surveys") {
  nd_all <- readRDS(here::here("data-generated/hbll-inside-grid.rds"))
  survey <- c("HBLL INS N", "HBLL INS S")
  type <- "HBLL (inside)"
  model_ssid <- c(39, 40)
}

if (region != "HBLL inside surveys") {
  all_data <- readRDS(here::here("data-raw/all_surv_catch.rds")) %>%
    filter(species_common_name == tolower(species))
} else {
    all_data <- readRDS(here::here("data-raw/all_surv_catch.rds")) %>%
    filter(stringr::str_detect(survey_abbrev, "HBLL INS")) %>% 
    filter(species_common_name == tolower(species))
}
```

Filter data 

```{r}
if (type == "Synoptic trawl") {
  d <- all_data %>%
    dplyr::filter(survey_abbrev %in% survey)
  units <- "density_kgpm2"
  d$density <- d[[units]] * 1000000 # change from per m2 to per km2
  family <- tweedie(link = "log") # to be used if not a delta model
} else {
  d <- all_data %>%
    dplyr::filter(survey_abbrev %in% survey)
  units <- "density_ppkm2"
  d$density <- d[[units]] # change from per m2 to per km2
  family <- tweedie(link = "log")
  # family <- nbinom2(link = 'log') # could use this instead, but would need to change units and add offset
}

if (region == "North (PMFC 5DE)") {
  d <- d %>% dplyr::filter(survey_abbrev == "SYN WCHG" | latitude > 53.3)
}

if (region == "South (PMFC 3CD5ABC)") {
  d <- d %>% dplyr::filter(survey_abbrev != "SYN WCHG" & latitude < 53.3)
}

if (region == "South 5 only (PMFC 5ABC)") {
  d <- d %>% dplyr::filter(survey_abbrev %in% c("SYN QCS", "SYN HS") & latitude < 53.3)
}

positive_sets <- dplyr::filter(d, density != 0)
prop_pos <- round(nrow(positive_sets) / nrow(d), digits = 3)
data <- d %>% dplyr::filter(survey_abbrev %in% unique(positive_sets$survey_abbrev))
ssid_string <- paste0(unique(data$survey_series_id), collapse = "n")
surv_string <- paste0(unique(data$survey_abbrev), collapse = ", ")
data <- sdmTMB::add_utm_columns(data, c("longitude", "latitude"), utm_crs = 32609)
```

Make mesh

```{r}
cutoff <- 20
if (region %in% c("HBLL outside surveys")) {
  cutoff <- 15
}
if (region %in% c("HBLL south outside survey", "HBLL north outside survey", "HBLL inside surveys", "QCS only", "HS only", "WCVI only")) {
  cutoff <- 10
}
if (region %in% c("WCHG only")) {
  cutoff <- 5
}

data <- data %>% mutate(present = ifelse(density > 0, 1, 0))
spde <- make_mesh(data, xy_cols = c("X", "Y"), cutoff = cutoff)

plot(spde)
```

```{r}
dir.create(here::here("data-generated/models"), showWarnings = FALSE)
```

Run sdmTMB model

```{r}
year_range <- range(data$year)
all_years <- data.frame(year = year_range[1]:year_range[2])
missing_years <- anti_join(all_years, data, by = "year") %>%
  select(year) %>%
  unique()
missing_years <- missing_years$year
if (length(missing_years) == 0L) missing_years <- NULL

tictoc::tic()
try({
  m1 <- sdmTMB(
    density ~ 1,
    data,
    mesh = spde,
    time = "year",
    extra_time = missing_years,
    family = delta_gamma(),
    spatial = "on",
    spatiotemporal = "RW",
    silent = params$silent,
    control = sdmTMBcontrol(newton_loops = 2L)
  )
})
tictoc::toc()
if (!exists("m1")) { m1 <- NULL}

saveRDS(m1, file = here::here(paste0("data-generated/models/m-", spp, "-", ssid_string, name, "-delta.rds")))
ok <- all(unlist(sanity(m1)))

if (!ok) {
  try({
    m1 <- sdmTMB(
      density ~ 1,
      data,
      mesh = spde,
      time = "year",
      extra_time = missing_years,
      family = delta_gamma(),
      spatial = list("on", "off"),
      spatiotemporal = "RW",
      silent = params$silent,
      control = sdmTMBcontrol(newton_loops = 2L)
    )
  })
  if (!exists("m1")) { m1 <- NULL}
  
  saveRDS(m1, file = here::here(paste0("data-generated/models/m-", spp, "-", ssid_string, name, "-delta.rds")))
  ok <- all(unlist(sanity(m1)))
}

if (!ok) {
  tictoc::tic()
  try({
    m <- sdmTMB(
      density ~ 1,
      data,
      mesh = spde,
      time = "year",
      extra_time = missing_years,
      family = family,
      spatial = "on",
      spatiotemporal = "RW",
      silent = params$silent,
      control = sdmTMBcontrol(newton_loops = 2L)
    )
  })
  tictoc::toc()
  if (!exists("m")) { m <- NULL}
  saveRDS(m, file = here::here(paste0("data-generated/models/m-", spp, "-", ssid_string, name, ".rds")))
} else {
  m <- m1
}
```

```{r}
p <- function(x) {
  out <- tryCatch(
    {
      print(x)
    },
    error = function(e) NULL
  )
  return(out)
}
p(m)
sanity(m)
```

Get index

```{r}
dir.create(here::here("data-generated/indices"), showWarnings = FALSE)

make_grid <- function(x, years) {
  .x <- filter(x, year == min(year)) # take one
  years <- sort(unique(years))
  .nd <- do.call(
    "rbind",
    replicate(length(years), .x, simplify = FALSE)
  )
  .nd[["year"]] <- rep(years, each = nrow(.x))
  .nd
}

converged <- function(fit) {
  all(unlist(sanity(fit)))
}

if (!is.null(m)) {
  if (converged(m)) {
    survey_type <- type
    # fitted_yrs <- sort(unique(m$data$year))
    yrs <- min(m$data$year):max(m$data$year)
    nd <- make_grid(nd_all, years = yrs)

    if (type == "Synoptic trawl") {
      nd <- filter(nd, survey %in% unique(positive_sets$survey_abbrev))
    } else {
      nd <- filter(nd, ssid %in% unique(positive_sets$survey_series_id))
    }
    nd <- na.omit(nd)
    nd$year <- as.integer(nd$year)

    samples <- data %>%
      count(year, survey_abbrev) %>%
      group_by(year) %>%
      mutate(freq = n / sum(n), total = sum(n)) %>%
      filter(n == max(n)) %>%
      select(year, dominant_surv = survey_abbrev, total, freq)
    
    p1 <- predict(m, newdata = nd, return_tmb_object = TRUE)
    i1 <- get_index(p1, area = nd$area, bias_correct = TRUE)
    i1 <- i1 %>% mutate(
      species = species,
      type = survey_type,
      region = region,
      surveys = paste(unique(positive_sets$survey_abbrev), collapse = ", "),
      ssids = paste(unique(positive_sets$survey_series_id), collapse = ", "),
      ssid_string = ssid_string,
      model = "delta-gamma",
      max_grad = max(m$gradients)
    )
    i1 <- left_join(i1, samples)
    saveRDS(i1, here::here(paste0("data-generated/indices/i-", spp, "-", ssid_string, name, ".rds")))
  }
}
```

```{r}
dir.create(here::here("figs", "spp-inds"), showWarnings = FALSE)

if (exists("i1")) {
  g <- ggplot(i1, aes(year, est / 1000)) +
    geom_line(colour = "red") +
    geom_ribbon(aes(ymin = lwr / 1000, ymax = upr / 1000), alpha = 0.2, fill = "red") +
    coord_cartesian(ylim = c(0, NA))
  g <- g + ylab("Biomass (tonnes)") +
    ggtitle(paste(species), subtitle = paste0(region, " (", surv_string))
  ggsave(here::here(paste0(here::here("figs", "spp-inds"), spp, "-", ssid_string, name, ".png")), height = 4, width = 6)
  g
}
```
